+++ teamscale-gradle-plugin/src/main/kotlin/com/teamscale/TeamscalePlugin.kt
@@ -1,11 +1,17 @@
 package com.teamscale
 
-import com.teamscale.config.TeamscalePluginExtension
+import com.teamscale.config.ReportConfigurationBase
+import com.teamscale.config.extension.TeamscaleJacocoReportTaskExtension
+import com.teamscale.config.extension.TeamscaleTestTaskExtension
+import com.teamscale.config.extension.TeamscalePluginExtension
 import org.gradle.api.GradleException
 import org.gradle.api.Plugin
 import org.gradle.api.Project
+import org.gradle.api.Task
 import org.gradle.api.plugins.JavaPlugin
+import org.gradle.api.tasks.testing.Test
 import org.gradle.testing.jacoco.plugins.JacocoPlugin
+import org.gradle.testing.jacoco.tasks.JacocoReport
 import org.gradle.util.GradleVersion
 
 
@@ -73,44 +79,90 @@ open class TeamscalePlugin : Plugin<Project> {
             extension = pluginExtension
         }
 
-        // Add the teamscale extension also to all test tasks
-        project.tasks.withType(TestImpacted::class.java) { testImpactedTask ->
-            configureTestImpactedTask(
-                project,
-                pluginExtension,
-                testImpactedTask
-            )
-        }
+        // Add the teamscale extension also to all TestImpacted tasks
+        extendTestImpactedTasks(project, pluginExtension)
+        extendTestTasks(project)
+        extendJacocoReportTasks(project)
     }
 
-
-    /** Configures the given impacted test executor. */
-    private fun configureTestImpactedTask(
+    private fun extendTestImpactedTasks(
         project: Project,
-        pluginExtension: TeamscalePluginExtension,
-        testImpacted: TestImpacted
+        pluginExtension: TeamscalePluginExtension
     ) {
-        project.logger.info("Configuring impacted tests executor task for ${project.name}:${testImpacted.name}")
-
-        val extension = pluginExtension.applyTo(testImpacted)
-
-        testImpacted.apply {
-            taskExtension = extension
-            dependsOn.add(project.configurations.getByName(impactedTestEngineConfiguration))
+        project.tasks.withType(TestImpacted::class.java) { testImpactedTask ->
+            /** Configures the given impacted test executor. */
+            project.logger.info(
+                "Configuring impacted tests executor task for ${project.name}:${
+                    testImpactedTask
+                        .name
+                }"
+            )
+            val extension = pluginExtension.applyTo(testImpactedTask)
+            testImpactedTask.apply {
+                taskExtension = extension
+                this.pluginExtension = pluginExtension
+                dependsOn.add(project.configurations.getByName(impactedTestEngineConfiguration))
+            }
+            val teamscaleReportTask = project.rootProject.tasks
+                .maybeCreate("${testImpactedTask.name}Report", TestwiseCoverageReportTask::class.java)
+            testImpactedTask.finalizedBy(teamscaleReportTask)
+            testImpactedTask.reportTask = teamscaleReportTask
+            teamscaleReportTask.apply {
+                testTaskName = testImpactedTask.name
+                configuration = extension
+            }
+            teamscaleReportTask.uploadTask = teamscaleUploadTask
+            registerReportAfterTask(testImpactedTask, extension.report)
         }
-
-        val teamscaleReportTask = project.rootProject.tasks
-            .maybeCreate("${testImpacted.name}Report", TeamscaleReportTask::class.java)
-        testImpacted.finalizedBy(teamscaleReportTask)
-
-        testImpacted.reportTask = teamscaleReportTask
-
-        teamscaleReportTask.apply {
-            testTaskName = testImpacted.name
-            configuration = extension
-        }
-
-        teamscaleReportTask.uploadTask = teamscaleUploadTask
     }
 
+    /**
+     * Adds an extension to the JacocoReport tasks and adds the report
+     * to the upload task when the extension has been configured.
+     */
+    private fun extendJacocoReportTasks(project: Project) {
+        project.tasks.withType(JacocoReport::class.java) { reportTask ->
+            teamscaleUploadTask.mustRunAfter(reportTask)
+
+            val extension =
+                reportTask.extensions.create(
+                    teamscaleExtensionName,
+                    TeamscaleJacocoReportTaskExtension::class.java,
+                    project,
+                    reportTask
+                )
+            registerReportAfterTask(reportTask, extension.report)
+        }
+    }
+
+    /**
+     * Adds an extension to the standard test tasks and adds the report
+     * to the upload task when the extension has been configured.
+     */
+    private fun extendTestTasks(project: Project) {
+        project.tasks.withType(Test::class.java) { testTask ->
+            if (testTask is TestImpacted) {
+                return@withType
+            }
+            val extension =
+                testTask.extensions.create(
+                    teamscaleExtensionName,
+                    TeamscaleTestTaskExtension::class.java,
+                    project,
+                    testTask
+                )
+            registerReportAfterTask(testTask, extension.report)
+        }
+    }
+
+    private fun registerReportAfterTask(
+        task: Task,
+        reportConfig: ReportConfigurationBase
+    ) {
+        task.doLast {
+            if (reportConfig.upload.get()) {
+                teamscaleUploadTask.reports.add(reportConfig.getReport())
+            }
+        }
+    }
 }