+++ teamscale-gradle-plugin/src/main/kotlin/com/teamscale/TeamscaleUploadTask.kt
@@ -1,7 +1,7 @@
 package com.teamscale
 
 import com.teamscale.client.TeamscaleClient
-import com.teamscale.config.TeamscalePluginExtension
+import com.teamscale.config.extension.TeamscalePluginExtension
 import org.gradle.api.DefaultTask
 import org.gradle.api.GradleException
 import org.gradle.api.tasks.Input
@@ -25,7 +25,15 @@ open class TeamscaleUploadTask : DefaultTask() {
     /** The commit for which the reports should be uploaded. */
     @get:Input
     val commitDescriptor
-        get() = extension.commit.getOrResolveCommitDescriptor(project)
+        get() = extension.commit.getOrResolveCommitDescriptor(project).first
+
+    /**
+     * The commit revision for which the reports should be uploaded.
+     * If set it is preferred over commitDescriptor.
+     */
+    @get:Input
+    val revision
+        get() = extension.commit.getOrResolveCommitDescriptor(project).second
 
     /** The list of reports to be uploaded. */
     @Input
@@ -50,7 +58,7 @@ open class TeamscaleUploadTask : DefaultTask() {
         server.validate()
 
         try {
-            logger.info("Uploading to $server at $commitDescriptor...")
+            logger.info("Uploading to $server at ${revision ?: commitDescriptor}...")
             uploadReports()
         } catch (e: Exception) {
             if (ignoreFailures) {
@@ -75,11 +83,18 @@ open class TeamscaleUploadTask : DefaultTask() {
             }
 
             try {
+                // Prefer to upload to revision and fallback to branch timestamp
+                val commitDescriptorOrNull = if (revision != null) null else commitDescriptor!!
                 retry(3) {
                     val client =
                         TeamscaleClient(server.url, server.userName, server.userAccessToken, server.project)
                     client.uploadReports(
-                        format, reportFiles, commitDescriptor, partition, "$message ($partition)"
+                        format,
+                        reportFiles,
+                        commitDescriptorOrNull,
+                        revision,
+                        partition,
+                        message
                     )
                 }
             } catch (e: ConnectException) {