+++ teamscale-gradle-plugin/src/main/kotlin/com/teamscale/config/JUnitReportConfiguration.kt
@@ -1,0 +1,13 @@
+package com.teamscale.config
+
+import com.teamscale.client.EReportFormat
+import org.gradle.api.Project
+import org.gradle.api.tasks.testing.Test
+
+/** Configuration for the JUnit results report. */
+open class JUnitReportConfiguration(project: Project, task: Test) :
+    ReportConfigurationBase(EReportFormat.JUNIT, project, task) {
+    init {
+        destination.set(project.providers.provider { task.reports.junitXml.destination })
+    }
+}